Messing around with templating/generics and the database queries
I was hoping to find a generic that did not require `impl`s but using an
`impl` does give the nice `.fn()` on objects and chaining, a pattern
common in Rust.

I also wanted to keep these outside of the `entities/` subfolder in case
`sea-orm` is ever used to regenerate our types after we change tables.