fix(sqlite): remove sqlite from daemon
In our testing, sqlite was too leaky of memory to perform well on a
MikroTik hAP ac lite. This caused frequent OOM kills, and short of
bundling a memory safe reimplementation of sqlite into the daemon by far
the easiest workaround for our goals is to instead attach a pi or
similar SBC to the AREDN routers and run postgres/daemon on those
instead.

Future considerations may be an RRD that works with the AREDN dashboard,
although removing tokio and swapping to a less heap heavy implementation
of reading packets from serial would be called for then.